checks: add checks for required annotations
This change adds a new check for a list of required annotations. For
now, only "charts.openshift.io/name" is checked to mitigate HELM-226
(see link below).

In terms of profile, the annotation check is enforced only in partner
and redhat profile for now.

Ref: https://issues.redhat.com/browse/HELM-226
Signed-off-by: Allen Bai <abai@redhat.com>

checks: update according to reviews

Signed-off-by: Allen Bai <abai@redhat.com>

profile: fix semver comparison

Signed-off-by: Allen Bai <abai@redhat.com>